ENH: Allow passing a COMMAND instead of reading from stdin
The initially intended processing of FILEs is not so useful; disregarding whitespace (and special files like /dev/null), a simple "test -s FILE" can already do the job. In any case, a "cat FILE | hasOutput" is fine. However, in some situations a synthesis of a single command can be useful, so let's accept commands instead.